+++ NEWS.md
@@ -61,6 +61,7 @@ ClimaAnalysis.replace(var, NaN => 0.0, missing => 0.0)
 if duplicate dates are detected.
 - Fix issue with plotting multiple figures at the same time.
 - Improve shading for `Visualize.heatmap2D_on_globe!`.
+- Add support for automatically converting CFTime.AbstractCFDateTime dates to seconds.
 
 v0.5.10
 -------